adding Authrization for edit and delete

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Question;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
@@ -78,7 +79,15 @@ class QuestionsController extends Controller
     public function edit(Question $question)
     {
         //
-       return view('questions.edit', compact('question'));
+        if(Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
+         return view('questions.edit', compact('question'));
+
+        // if(Gate::allows('update-question', $question)){
+        //     return view('questions.edit', compact('question'));
+        // }
+        // abort(403, "Access denied");
     }
 
     /**
@@ -91,6 +100,10 @@ class QuestionsController extends Controller
     public function update(AskQuestionRequest $request, Question $question)
     {
         //
+        if(Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
+         return view('questions.edit', compact('question'));
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', 'Your question has been updated');
     }
@@ -104,7 +117,10 @@ class QuestionsController extends Controller
     public function destroy(Question $question)
     {
         //
-        $question->delete();
+        if(Gate::denies('delete-question', $question)){
+            abort(403, "Access denied");
+        }
+         $question->delete();
         return redirect('/questions')->with('success', 'Your question has been deleted');
     }
 }
